requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
ping3==2.6.5
prettytable==0.7.2
python-engineio==3.13.0
python-socketio==4.6.0
requests==2.23.0
selenium==3.141.0
six==1.15.0
tcping==0.1.1rc1
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
